ABLASTR: Move IntervalsParser
Move the `IntervalsParser` for time intervals to ABLASTR, so we
can reuse it in other codes. I need it next in ImpactX :)
