+++ Source/ablastr/utils/text/CMakeLists.txt
@@ -2,6 +2,7 @@ foreach(D IN LISTS WarpX_DIMS)
     warpx_set_suffix_dims(SD ${D})
     target_sources(ablastr_${SD}
       PRIVATE
+        IntervalsParser.cpp
         StreamUtils.cpp
         StringUtils.cpp
     )
